check own

+++ src/employee/employee.controller.ts
@@ -7,6 +7,7 @@ import {
   Param,
   Delete,
   Query,
+  Headers,
 } from "@nestjs/common";
 import { EmployeeService } from "./employee.service";
 import { EmployeeEntity } from "./entity/employee.entity";
@@ -17,18 +18,24 @@ import { Common } from "./../../helper/common/common";
 import { UserEntity } from "src/user/entity/user.entity";
 import { UserService } from "src/user/user.service";
 import { DeleteResult } from "typeorm";
+import { JWTUtil } from "src/auth/JWTUtil";
 
 @Controller("employee")
 export class EmployeeController {
   constructor(
     private readonly service: EmployeeService,
-    private readonly userService: UserService
+    private readonly userService: UserService,
+    private readonly jwtUtil: JWTUtil,
   ) {}
 
   @Post()
   async create(@Body() body): Promise<ApiResponse<EmployeeEntity>> {
     try {
       if (await Common.verifyRequest(body.cksRequest, body.timeRequest)) {
+        const employee = await this.service.findByPhone(body.phone);
+        if (employee) {
+          return ResponseHelper.error(0, "Số điện thoại đã được sử dụng");
+        }
         const user = await this.userService.findById(body.store_id);
         if (user) {
           body.keySearch =
@@ -50,9 +57,13 @@ export class EmployeeController {
   async findAll(
     @Query("page") page: number = 1,
     @Query("limit") limit: number = 10,
-    @Query() query
+    @Query() query,
+    @Headers('Authorization') auth: string
   ): Promise<ApiResponse<EmployeeEntity[]>> {
     try {
+      
+      const json = await this.jwtUtil.decode(auth);
+      console.log(json.shop)
       if (await Common.verifyRequest(query.cksRequest, query.timeRequest)) {
         const [res, totalCount] = await this.service.findAll(
           page,
@@ -93,6 +104,11 @@ export class EmployeeController {
   async update(@Body() body: any): Promise<ApiResponse<UpdateResult>> {
     try {
       if (await Common.verifyRequest(body.cksRequest, body.timeRequest)) {
+        const employee = await this.service.findOne(body.id)
+        const store_id = await Common.getIdShop(body.cksRequest)
+        if (!employee && store_id != employee.store_id) {
+          return ResponseHelper.error(0, "Lỗi");
+        }
         delete body["cksRequest"];
         delete body["timeRequest"];
         const res = await this.service.update(body);
@@ -107,6 +123,11 @@ export class EmployeeController {
   async remove(@Param() param, @Query() query) {
     try {
       if (await Common.verifyRequest(query.cksRequest, query.timeRequest)) {
+        const employee = await this.service.findOne(param.id)
+        const store_id = await Common.getIdShop(query.cksRequest)
+        if (store_id != employee.store_id) {
+          return ResponseHelper.error(0, "Lỗi");
+        }
         const res = await this.service.remove(param.id);
         return ResponseHelper.success(res);
       }
@@ -114,6 +135,4 @@ export class EmployeeController {
       return ResponseHelper.error(0, error);
     }
   }
-
-
 }
